Initial Commit
These are all of the files I used in making my collages. The folder
"collage" has all the code that I call in the "makecollage.py" file. To
use it, paste the "collage" folder in somewhere that your python PATH
can see, and then run the code in the  "makecollage.py" file in python
(preferably an IDE like Spyder). The output will be all the stuff you
need to paste into MS Paint or Photoshop and play around with.